+++ src/main/java/io/camunda/tasklist/ProcessApplication.java
@@ -4,10 +4,12 @@ import io.camunda.zeebe.spring.client.EnableZeebeClient;
 import io.camunda.zeebe.spring.client.annotation.Deployment;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 @EnableZeebeClient
 @Deployment(resources = "classpath*:/bpmn/*.*")
+@ComponentScan({"io.camunda.tasklist", "io.camunda.operate"})
 public class ProcessApplication {
 
   public static void main(String[] args) {