+++ src/main/java/io/camunda/tasklist/facade/ProcessController.java
@@ -1,7 +1,12 @@
 package io.camunda.tasklist.facade;
 
-import io.camunda.tasklist.ProcessVariables;
+import io.camunda.operate.rest.OperateRestClient;
+import io.camunda.operate.rest.dto.ProcessDefinitionQuery;
+import io.camunda.operate.rest.dto.ProcessDefinitionQueryResults;
+import io.camunda.operate.rest.exception.OperateException;
+import io.camunda.operate.rest.exception.OperateRestException;
 import io.camunda.zeebe.client.ZeebeClient;
+import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,15 +18,18 @@ import org.springframework.web.bind.annotation.*;
 public class ProcessController {
 
   private static final Logger LOG = LoggerFactory.getLogger(ProcessController.class);
-  private final ZeebeClient zeebe;
+  final ZeebeClient zeebe;
+  final OperateRestClient operateRestClient;
 
-  public ProcessController(@Autowired ZeebeClient client) {
+  public ProcessController(
+      @Autowired ZeebeClient client, @Autowired OperateRestClient operateRestClient) {
     this.zeebe = client;
+    this.operateRestClient = operateRestClient;
   }
 
   @PostMapping("/start/{processId}")
   public void startProcessInstance(
-      @PathVariable String processId, @RequestBody ProcessVariables variables) {
+      @PathVariable String processId, @RequestBody Map<String, Object> variables) {
 
     LOG.info("Starting process `" + processId + "` with variables: " + variables);
 
@@ -37,7 +45,7 @@ public class ProcessController {
   public void publishMessage(
       @PathVariable String messageName,
       @PathVariable String correlationKey,
-      @RequestBody ProcessVariables variables) {
+      @RequestBody Map<String, Object> variables) {
 
     LOG.info(
         "Publishing message `{}` with correlation key `{}` and variables: {}",
@@ -51,5 +59,14 @@ public class ProcessController {
         .correlationKey(correlationKey)
         .variables(variables)
         .send();
+  }
+
+  @PostMapping("/process-definitions/search")
+  public ProcessDefinitionQueryResults processDefinitionSearch(
+      @RequestBody ProcessDefinitionQuery query) throws OperateRestException, OperateException {
+
+    LOG.info("Searching for process definition `{}`", query);
+
+    return operateRestClient.query(query);
   }
 }