+++ src/main/js/src/App.js
@@ -1,6 +1,8 @@
 import React, { Component } from 'react';
 import './app.scss'
-import WelcomeForm from "./components/WelcomeForm";
+import CamundaForm from "./components/CamundaForm";
+
+const welcomeFormSchema = require('./forms/welcome.form.json')
 
 let rest = require('rest');
 let mime = require('rest/interceptor/mime');
@@ -8,6 +10,7 @@ let client = rest.wrap(mime);
 
 const restApi = "http://localhost:8087";
 const processId = "Process_twoUserTasks";
+const pollingIntervalMillis = 500;
 
 const welcomeFormData = {
   "firstName": "Dave",
@@ -27,7 +30,6 @@ const initial = {
   schema: null,
   data: welcomeFormData,
   task: null,
-  user: { userName: null },
   tasks: [],
   screen: 'welcome',
   history: {}
@@ -38,7 +40,16 @@ class App extends Component {
     super(props);
     this.state = initial;
 
+    const query = new URLSearchParams(window.location.search);
+    const userName = query.get('userName');
+    if(userName) {
+      this.state.data.userName = userName;
+      this.state.screen = 'searching';
+    }
+
+    this.poll = this.poll.bind(this);
     this.createInstance = this.createInstance.bind(this);
+    this.getAssignedTasks = this.getAssignedTasks.bind(this);
     this.onFormSubmit = this.onFormSubmit.bind(this);
 
   }
@@ -58,24 +69,64 @@ class App extends Component {
       console.log("form submitted ...");
       console.log(data);
       this.setState({data: data});
-      this.createInstance(data)
+      this.createInstance(data);
+      this.state.screen = 'searching';
     } else {
       console.log("form submitted but has errors ...");
       console.log(errors);
     }
   }
 
+  getAssignedTasks(assignee) {
+    return client({
+      path: `${restApi}/tasks/search`,
+      headers: {'Content-Type': 'application/json'},
+      entity: {
+        "state": "CREATED",
+        "assigned": true,
+        "assignee": `${assignee}`
+      }
+    })
+  }
+
+  poll() {
+    //console.log('polling ...');
+    if(this.state.data.userName && this.state.screen === "searching") {
+
+      console.log("check for new tasks ...");
+      this.getAssignedTasks(this.state.data.userName)
+        .then((response) => {
+          let tasks = response.entity;
+          if(tasks && tasks.length > 0) {
+            this.setState({
+              task: tasks[0],
+              history: {lastCheck: new Date()},
+              screen: 'loadForm'
+              //TODO: find form based on tasks[0].formKey
+            });
+          }
+          this.setState({history: {lastCheck: new Date()}});
+        });
+
+    }
+  }
+
+  componentDidMount() {
+    this.timerId = setInterval(() => this.poll(), pollingIntervalMillis);
+  }
+
   render() {
     if(this.state.screen === 'welcome') {
       return (
-        <WelcomeForm
+        <CamundaForm
+          schema={welcomeFormSchema}
           data={this.state.data}
           onSubmit={this.onFormSubmit}
         />
       );
     } else {
       // Default
-      return (<div>Error</div>);
+      return (<div>Current State: {this.state.screen}</div>);
     }
   }
 }