+++ src/main/js/src/components/CamundaForm/index.js
@@ -1,0 +1,2 @@
+import CamundaForm from './CamundaForm';
+export default CamundaForm;