+++ src/main/js/src/components/WelcomeForm/index.js
@@ -1,2 +1,0 @@
-import WelcomeForm from './WelcomeForm';
-export default WelcomeForm;